Create IsVisitedAndRevisedAfterConstraintStrategy
